Lr5

+++ backend/src/controllers/UserController.ts
@@ -2,9 +2,10 @@ import { NextFunction, Request, Response } from 'express';
 import UserService from '@services/UserService.js';
 import { Roles } from '@constants/Roles.js';
 import CustomError from '@utils/CustomError.js';
-import { ErrorCodes } from '@constants/Errors.js';
-import { ChangeUserRoleDto } from '@dto/UserDto.js';
+import { ErrorCodes, ErrorMessages } from '@constants/Errors.js';
 import UserMapper from '@mappers/UserMapper.js';
+import User from '@models/User.js';
+import { roleSchema, UpdateUserInput, updateUserSchema, idSchema } from '@validation/user.js';
 
 class UserController {
     async getAllUsers(req: Request, res: Response, next: NextFunction) {
@@ -20,12 +21,34 @@ class UserController {
 
     async getUser(req: Request, res: Response, next: NextFunction) {
         try {
-            const id = Number(req.params.id);
+            const requester = req.user as User;
+            const id = idSchema.parse(req.params.id);
+            let sendDeletedEvents = false;
 
-            UserController.validateUserId(id);
-
-            const user = await UserService.getUser(id);
+            if (requester.id === id || requester.role === 'admin') {
+                sendDeletedEvents = true;
+            }
+            const user = await UserService.getUser(id, sendDeletedEvents);
             res.status(200).json(UserMapper.toResponseDto(user));
+        } catch (e) {
+            next(e);
+        }
+    }
+
+    async updateUser(req: Request, res: Response, next: NextFunction) {
+        try {
+            const currentUser = req.user as User;
+            const targetUserId = idSchema.parse(req.params.id);
+            const userData: UpdateUserInput = updateUserSchema.parse(req.body);
+            if (currentUser.id !== targetUserId && currentUser.role !== 'admin') {
+                return next(new CustomError(ErrorCodes.ForbiddenError, ErrorMessages.ForbiddenError));
+            }
+            if (currentUser.role !== 'admin' && userData.role && userData.role !== currentUser.role) {
+                return next(new CustomError(ErrorCodes.ForbiddenError, 'Вы не можете изменить свою роль.'));
+            }
+
+            const updatedUser = await UserService.updateUser(targetUserId, userData);
+            res.status(200).json(UserMapper.toResponseDto(updatedUser));
         } catch (e) {
             next(e);
         }
@@ -33,11 +56,13 @@ class UserController {
 
     async deleteUser(req: Request, res: Response, next: NextFunction) {
         try {
-            const id = Number(req.params.id);
+            const requester = req.user as User;
+            const id = idSchema.parse(req.params.id);
             const hardDelete = req.query.hardDelete === 'true' || req.query.hardDelete === '1';
 
-            UserController.validateUserId(id);
-
+            if (requester.id !== id || requester.role !== 'admin') {
+                return next(new CustomError(ErrorCodes.ForbiddenError, ErrorMessages.ForbiddenError));
+            }
             const result = await UserService.deleteUser(id, hardDelete);
             res.status(200).json(result);
         } catch (e) {
@@ -47,9 +72,12 @@ class UserController {
 
     async restoreUser(req: Request, res: Response, next: NextFunction) {
         try {
-            const id = Number(req.params.id);
+            const requester = req.user as User;
+            const id = idSchema.parse(req.params.id);
 
-            UserController.validateUserId(id);
+            if (requester.id !== id || requester.role !== 'admin') {
+                return next(new CustomError(ErrorCodes.ForbiddenError, ErrorMessages.ForbiddenError));
+            }
 
             const result = await UserService.restoreUser(id);
             res.status(200).json(result);
@@ -60,9 +88,7 @@ class UserController {
 
     async getUserRole(req: Request, res: Response, next: NextFunction) {
         try {
-            const id = Number(req.params.id);
-
-            UserController.validateUserId(id);
+            const id = idSchema.parse(req.params.id);
 
             const role = await UserService.getUserRole(id);
             res.status(200).json(role);
@@ -72,31 +98,23 @@ class UserController {
     }
 
     async setUserRole(req: Request, res: Response, next: NextFunction): Promise<void> {
-        const id = Number(req.params.id);
-        const { role } = req.body as ChangeUserRoleDto;
-
         try {
-            UserController.validateUserId(id);
+            const id = idSchema.parse(req.params.id);
+            const role = roleSchema.parse(req.body);
+
+            if (!role || !Object.values(Roles).includes(role)) {
+                return next(
+                    new CustomError(
+                        ErrorCodes.BadRequest,
+                        `Invalid or missing role. Allowed roles are: ${Object.values(Roles).join(', ')}`,
+                    ),
+                );
+            }
+
+            const user = await UserService.setUserRole(id, role);
+            res.status(200).json(user);
         } catch (e) {
             next(e);
-        }
-
-        if (!role || !Object.values(Roles).includes(role)) {
-            return next(
-                new CustomError(
-                    ErrorCodes.BadRequest,
-                    `Invalid or missing role. Allowed roles are: ${Object.values(Roles).join(', ')}`,
-                ),
-            );
-        }
-
-        const user = await UserService.setUserRole(id, role);
-        res.status(200).json(user);
-    }
-
-    private static validateUserId(id: number): void {
-        if (!Number.isInteger(id) || id <= 0) {
-            throw new CustomError(ErrorCodes.BadRequest, 'Invalid user ID. It must be a positive integer.');
         }
     }
 }
